% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancom2.R
\name{ANCOM}
\alias{ANCOM}
\title{Analysis of Composition of Microbiomes}
\usage{
ANCOM(
  feature_table,
  meta_data,
  struc_zero = NULL,
  main_var,
  p_adj_method = "BH",
  alpha = 0.05,
  adj_formula = NULL,
  rand_formula = NULL,
  paired,
  pid,
  count,
  ...
)
}
\arguments{
\item{feature_table}{, object from the feature function}

\item{meta_data}{, object from the feature function}

\item{struc_zero}{, object from the feature function}

\item{main_var}{, group to compare}

\item{p_adj_method}{, defalut BH}

\item{alpha}{,defalut 0.05}

\item{paired}{, T/F}

\item{pid, }{patient ID}

\item{...}{}
}
\value{

}
\description{
Analysis of Composition of Microbiomes
}
