% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureMeta.R
\name{centorComp}
\alias{centorComp}
\title{centroComp
compare the center distance between groups}
\usage{
centorComp(pro, method, config, color)
}
\arguments{
\item{pro}{metagenome profile ,row is sample ID}

\item{method}{distance method}

\item{config}{group info}

\item{color}{set group color in plot}
}
\value{

}
\description{
centroComp
compare the center distance between groups
}
