% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statMeta.R
\name{checkOutlier}
\alias{checkOutlier}
\title{checkOutlier
output the outlier value on a vector by quantile}
\usage{
checkOutlier(x, coef = 2, sname, plot = F, vname = "test")
}
\arguments{
\item{vname}{}
}
\description{
checkOutlier
output the outlier value on a vector by quantile
}
