% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statNew.R
\name{clusterTrajector}
\alias{clusterTrajector}
\title{clusterTrajector}
\usage{
clusterTrajector(dataset, phemeta, span)
}
\arguments{
\item{dataset}{metabolism}

\item{phemeta}{clinical paremeter eg, age}

\item{span}{loess paremeter}
}
\value{

}
\description{
clusterTrajector
}
