% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMeta.R
\name{comparePair}
\alias{comparePair}
\title{comparePair
wilcox pair test for multi time point}
\usage{
comparePair(data, config, group, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
comparePair
wilcox pair test for multi time point
}
