% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureMeta.R
\name{corPlot}
\alias{corPlot}
\title{corPlot
pheatmap to figure the correlation}
\usage{
corPlot(corres, cutoff, adjust, tr)
}
\arguments{
\item{tr}{}
}
\value{

}
\description{
corPlot
pheatmap to figure the correlation
}
