% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMeta.R
\name{cumNormStatFast2}
\alias{cumNormStatFast2}
\title{cum Norm stat method form ANCOM_BC}
\usage{
cumNormStatFast2(data, pFlag = FALSE, rel = 0.1, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
cum Norm stat method form ANCOM_BC
}
