% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMeta.R
\name{CvLassomulti}
\alias{CvLassomulti}
\title{CvLasso
selsect feature on diffrent group}
\usage{
CvLassomulti(metadata, dataset, response, feturetop = NULL)
}
\arguments{
\item{metadata}{, response}

\item{dataset}{, metabolism or microbe}

\item{feturetop}{}
}
\value{

}
\description{
CvLasso
selsect feature on diffrent group
}
