% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statMeta.R
\name{distJSD}
\alias{distJSD}
\title{JSD distance
to compute the jsd distance}
\usage{
distJSD(inMatrix, pseudocount = 1e-08, ...)
}
\arguments{
\item{...}{}
}
\value{
distance
}
\description{
JSD distance
to compute the jsd distance
}
