% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMeta.R
\name{elibuqMeta}
\alias{elibuqMeta}
\title{the effective library size using UQ (upper quartile)}
\usage{
elibuqMeta(data, p = 0.75)
}
\arguments{
\item{p}{}
}
\value{

}
\description{
the effective library size using UQ (upper quartile)
}
