% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancom2.R
\name{feature_table_pre_process}
\alias{feature_table_pre_process}
\title{feature_table_pre_process}
\usage{
feature_table_pre_process(
  feature_table,
  meta_data,
  sample_var,
  group_var = NULL,
  out_cut = 0.05,
  zero_cut = 0.9,
  lib_cut,
  neg_lb,
  count = F
)
}
\arguments{
\item{feature_table}{,feature table should be a matrix/data.frame with each feature in rows and sample in columns.}

\item{meta_data, }{Metadata should be a matrix/data.frame containing the sample identifier.}

\item{sample_var, }{sample ID name}

\item{group_var, }{the group to compare}

\item{out_cut, }{alpha}

\item{zero_cut, }{occurrance rate}

\item{lib_cut, }{library size}

\item{neg_lb, }{logistic}

\item{count, }{logistic}
}
\value{

}
\description{
This function is from
\href{https://github.com/FrederickHuangLin/ANCOM}{ANCOM}
}
