% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statMeta.R
\name{filterPer}
\alias{filterPer}
\title{filterPer}
\usage{
filterPer(x, row, percent, include = T)
}
\arguments{
\item{x}{, data.frame}

\item{row}{, row or col}

\item{percent}{, ratio}

\item{include}{, belong or exclude}
}
\value{
dataframe
}
\description{
filterPer
}
