% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicsMeta.R
\name{getMetahlan2rank}
\alias{getMetahlan2rank}
\title{getMetaphlan2rank
get the phylum-species rank}
\usage{
getMetahlan2rank(name)
}
\arguments{
\item{name}{, the merge metaphlan2 rownames}
}
\value{

}
\description{
getMetaphlan2rank
get the phylum-species rank
}
