% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureMeta.R
\name{ggcorr}
\alias{ggcorr}
\title{correlation between pairwise}
\usage{
ggcorr(
  data,
  method = c("pairwise", "pearson"),
  cor_matrix = NULL,
  nbreaks = NULL,
  digits = 2,
  name = "",
  low = "#3B9AB2",
  mid = "#EEEEEE",
  high = "#F21A00",
  midpoint = 0,
  palette = NULL,
  geom = "tile",
  min_size = 2,
  max_size = 6,
  label = FALSE,
  label_alpha = FALSE,
  label_color = "black",
  label_round = 1,
  label_size = 4,
  limits = TRUE,
  drop = !limits,
  layout.exp = 0,
  legend.position = "right",
  legend.size = 9,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
correlation between pairwise
}
