% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statMeta.R
\name{kBest}
\alias{kBest}
\title{kBest
to get the best k in non-supervision method}
\usage{
kBest(data, dist, method = "kmeans")
}
\arguments{
\item{method}{}
}
\value{
list , include the best cluster number k & original figure & dataframe
}
\description{
kBest
to get the best k in non-supervision method
}
