% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{log-job}
\alias{log-job}
\alias{log-job}
\alias{read_job_executions}
\alias{job_start}
\alias{job_end}
\alias{job_error}
\title{Metadata for task execution}
\usage{
read_job_executions(job = "", days = 7L)

job_start(job)

job_end(job)

job_error(job)
}
\arguments{
\item{job}{name of the data pipeline}

\item{destination}{name of the table that data is written to}

\item{records}{number of records processed by the job}

\item{increment}{the largest value in the increment field of the destination}
}
\description{
Functions that allow to log and query job executions.
}
\details{
`read_job_executions` - reads execution log for a given job

`log_job_start` - logs start of the job execution

`log_job_end` - logs end of the job exectuion

`log_job_error` - logs errored job
}
