% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{log-model}
\alias{log-model}
\alias{log_model_performance}
\title{Logs performance of the statistical or machine learning model into InfluxDb}
\usage{
log_model_performance(dataset, target, test, algorithm, metric, group, records,
  value)
}
\arguments{
\item{dataset}{name of the dataset that was targed by the model, e.g. `sessions`}

\item{target}{explained variable, e.g. `conversion rate``}

\item{test}{boolean defines whether validation was done on test or training dataset}

\item{algorithm}{name of the method that was used to create the model}

\item{metric}{name of the model performance metric, e.g. rsquared or mape}

\item{group}{group within the data if data is split into different segments}

\item{records}{size of the sample that was used to estimate the model}

\item{value}{value of the model performance metric on a given sample}
}
\description{
Logs performance of the statistical or machine learning model into InfluxDb
}
