% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMeta.R
\name{matchpairID}
\alias{matchpairID}
\title{matchpair ID
to match the sample ID for pair wilcox test}
\usage{
matchpairID(configdat, ID, Time, num = 2)
}
\arguments{
\item{configdat}{, dataframe}

\item{ID}{, char: sample ID}

\item{Time}{, char: time point}

\item{num, }{numberic}
}
\value{
dataframe
}
\description{
matchpair ID
to match the sample ID for pair wilcox test
}
