% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textMeta.R
\name{mergeP}
\alias{mergeP}
\title{mergeP
merge two dataframe ,who have diffrent colname & diffrent rowname}
\usage{
mergeP(dat1, dat2)
}
\arguments{
\item{by}{}
}
\value{

}
\description{
mergeP
merge two dataframe ,who have diffrent colname & diffrent rowname
}
