% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{multiAdonis}
\alias{multiAdonis}
\title{multiAdonis
for pair compare for multigroup use the adonis}
\usage{
multiAdonis(dat, config, perm = 999, method = "bray")
}
\arguments{
\item{method}{}
}
\value{
data.frame
}
\description{
multiAdonis
for pair compare for multigroup use the adonis
}
