% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureMeta.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot
this function to plot volcano based ggplot2}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{layout}{}
}
\description{
Multiplot
this function to plot volcano based ggplot2
}
