% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statMeta.R
\name{mutivarOutlier}
\alias{mutivarOutlier}
\title{mutivarOutlier
on a dataframe to select the outlier}
\usage{
mutivarOutlier(dat, scale = F, naaction = F, k, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
mutivarOutlier
on a dataframe to select the outlier
}
