% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicsMeta.R
\name{myGeeAnalysis}
\alias{myGeeAnalysis}
\title{myGeeAnalysis}
\usage{
myGeeAnalysis(dataset, metadata, confounder, timevar, scale, IDvar, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
myGeeAnalysis
}
