% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{myvolcano}
\alias{myvolcano}
\title{plot volcano
this function to plot volcano based ggplot2}
\usage{
myvolcano(qdat, fdcut)
}
\arguments{
\item{qdat}{include column name Fdr & fc}

\item{fdcut}{the foldchange cutoff}
}
\value{
figure
}
\description{
plot volcano
this function to plot volcano based ggplot2
}
