% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMeta.R
\name{mywilcox_2t}
\alias{mywilcox_2t}
\title{mywilcox_2t
wilcox pair test for 2 time point}
\usage{
mywilcox_2t(datamatrix, configdata, Time2, ratio = "zero", ...)
}
\arguments{
\item{datamatrix}{,}

\item{configdata}{, from function matchpairID}

\item{Time2}{, time point}

\item{ratio}{, use NA or 0 to compute the ratio}

\item{...}{}
}
\value{

}
\description{
mywilcox_2t
wilcox pair test for 2 time point
}
