% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureMeta.R
\name{nmdsFigEx}
\alias{nmdsFigEx}
\title{nmdsFigEx
plot the nmds figure}
\usage{
nmdsFigEx(data, method = "bray", config, color = c("#66C2A5", "#E78AC3"))
}
\arguments{
\item{color}{}
}
\value{

}
\description{
nmdsFigEx
plot the nmds figure
}
