% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkMeta.R
\name{pairNet}
\alias{pairNet}
\title{pairNet}
\usage{
pairNet(Cor1, Cor2, cutoff, species)
}
\arguments{
\item{Cor1}{, corelation matrix 1}

\item{Cor2}{, corelation matrix 2}

\item{cutoff, }{the eage cutoff}

\item{species, }{retain the species}
}
\description{
pairNet
}
