% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMeta.R
\name{r2Towpartmodel}
\alias{r2Towpartmodel}
\title{r2Twopartmodel}
\usage{
r2Towpartmodel(dat, phe, response, cutoff, number = 10, cutoffp = 0.01,
  repeatN = 100, confunder = F)
}
\arguments{
\item{dat}{, microbiome data row is sample id, col is variable}

\item{phe}{, metadata row is sample id ,col is variable}

\item{response}{, the variable to expain}

\item{cutoff}{, ensure  to detect or undetect}

\item{number}{, when sample number is limited , default 10}

\item{cutoffp}{, significant feature}

\item{repeatN}{, repeat number}
}
\value{
vector
}
\description{
From The Gut Microbiome Contributes to a Substantial Proportion of the Variation in Blood Lipids
}
