% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMeta.R
\name{r2Twopartmodelcv}
\alias{r2Twopartmodelcv}
\title{r2Twopartmodelcv
add the cross validation process
From The Gut Microbiome Contributes to a Substantial Proportion of the Variation in Blood Lipids}
\usage{
r2Twopartmodelcv(
  dat,
  phe,
  response,
  cutoff,
  number = 10,
  cutoffp = 0.01,
  repeatN = 100,
  fold,
  confounder = NULL
)
}
\arguments{
\item{dat}{, microbiome data row is sample id, col is variable}

\item{phe}{, metadata row is sample id ,col is variable}

\item{response}{, the variable to expain}

\item{cutoff}{, ensure  to detect or undetect}

\item{number}{, when sample number is limited , default 10}

\item{cutoffp}{, significant feature}

\item{repeatN}{, repeat number}

\item{fold}{, cross validation fold}
}
\value{
vector
}
\description{
r2Twopartmodelcv
add the cross validation process
From The Gut Microbiome Contributes to a Substantial Proportion of the Variation in Blood Lipids
}
