% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMeta.R
\name{RankINT}
\alias{RankINT}
\title{Based-rank Inverse Normal Transformation
The detail inf is (here)[https://cran.r-project.org/web/packages/RNOmni/vignettes/RNOmni.html]}
\usage{
RankINT(u, k = 0.375)
}
\arguments{
\item{u, }{vector}

\item{k, }{0<= k <=0.5}
}
\value{
out, vector transformed
}
\description{
Based-rank Inverse Normal Transformation
The detail inf is (here)[https://cran.r-project.org/web/packages/RNOmni/vignettes/RNOmni.html]
}
\examples{
x <- c(0.1, 0, 3, 0.5, NA)
RankINT(x)
}
