% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkMeta.R
\name{renorm}
\alias{renorm}
\title{renorm
normalization the compsition data}
\usage{
renorm(dat)
}
\arguments{
\item{dat}{, dataframe row is variable , col is sample ID}
}
\value{
dataframe
}
\description{
renorm
normalization the compsition data
}
\examples{

}
