% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMeta.R
\name{smote.exs2}
\alias{smote.exs2}
\title{smote.exs2
this function from package DMwR smote.exs}
\usage{
smote.exs2(data, tgt, N, k)
}
\arguments{
\item{k}{}
}
\value{

}
\description{
smote.exs2
this function from package DMwR smote.exs
}
