% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkMeta.R
\name{sparccNet}
\alias{sparccNet}
\title{sparccNet}
\usage{
sparccNet(dat, cutoff, main, layout = "layout.circle")
}
\arguments{
\item{dat}{, dataframe  rawdata}

\item{cutoff}{,numberic 0.3}

\item{main}{, str topic of figure}

\item{layout}{, the lay style of node layout}
}
\value{
figure
}
\description{
sparccNet
}
