% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicsMeta.R
\name{splitMetaphlan2}
\alias{splitMetaphlan2}
\title{split the metaphlan2 dataset}
\usage{
splitMetaphlan2(data, prefix)
}
\arguments{
\item{data}{, merge metaphlan2 dataset}

\item{prefix}{, output name}
}
\value{
list, include all rank data
}
\description{
split the metaphlan2 dataset
}
