% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureMeta.R
\name{topTax}
\alias{topTax}
\title{topTax
plot the high abundance tax}
\usage{
topTax(metadata, K = 20, rmUnclass = F)
}
\arguments{
\item{metadata}{, tax profile row is tax col is sample}

\item{K}{, top number}

\item{rmUnclass}{, if to remove the unclass tax}
}
\value{
figure
}
\description{
topTax
plot the high abundance tax
}
