% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statNew.R
\name{twopartEta}
\alias{twopartEta}
\title{twopartEta
to compute the partial eta or relative risk}
\usage{
twopartEta(dat, phe, number = 20)
}
\arguments{
\item{number}{}
}
\value{

}
\description{
twopartEta
to compute the partial eta or relative risk
}
