% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMeta.R
\name{twopartModel}
\alias{twopartModel}
\title{twopartModel
From The Gut Microbiome Contributes to a Substantial Proportion of the Variation in Blood Lipids}
\usage{
twopartModel(dat, phe, response, cutoff, number = 10)
}
\arguments{
\item{dat}{, microbiome data row is sample id, col is variable}

\item{phe}{, metadata row is sample id ,col is variable}

\item{response}{, the variable to explain}

\item{cutoff}{, ensure to detect or undetect}

\item{number}{, when sample number is limited , default 10}
}
\value{
dataframe
}
\description{
twopartModel
From The Gut Microbiome Contributes to a Substantial Proportion of the Variation in Blood Lipids
}
