% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancom2.R
\name{feature_table_pre_process}
\alias{feature_table_pre_process}
\title{feature_table_pre_process

This function is from [here] (https://github.com/FrederickHuangLin/ANCOM)}
\usage{
feature_table_pre_process(
  feature.table,
  meta.data,
  sample.var,
  group.var,
  zero.cut = 0.9,
  lib.cut = 1000,
  neg.lb
)
}
\arguments{
\item{feature_table}{,feature table should be a matrix/data.frame with each feature in rows and sample in columns.}

\item{meta_data, }{Metadata should be a matrix/data.frame containing the sample identifier.}

\item{sample_var, }{sample ID name}

\item{group_var, }{the group to compare}

\item{zero_cut, }{occurrance rate}

\item{lib_cut, }{library size}

\item{neg_lb}{}
}
\value{

}
\description{
feature_table_pre_process

This function is from [here] (https://github.com/FrederickHuangLin/ANCOM)
}
