% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{influxConnection}
\alias{influxConnection}
\title{Connection to the InfluxDb using environtment variables}
\usage{
influxConnection(scheme = c("http", "https"),
  host = Sys.getenv("INFLUX_HOST"), user = Sys.getenv("INFLUX_USERNAME"),
  port = Sys.getenv("INFLUX_PORT", 8086),
  pwd = Sys.getenv("INFLUX_PASSWORD"))
}
\description{
Requires:
 INFLUX_HOST
 INFLUX_USERNAME
 INFLUX_PORT
 INFLUX_PASSWORD
}
